requirements.txt
aiogram==3.21.0
sqlalchemy==2.0.22
asyncpg==0.30.0
python-dotenv==1.0.0
APScheduler==3.11.0